switched to credentials based login

+++ pages/api/allow/remove.ts
@@ -16,10 +16,10 @@ handler.post(async (req: NextApiRequest, res: NextApiResponse) => {
 			? [...req.body.ids]
 			: [];
 		for (var i = 0; i < ids.length; i++) {
-			let contact = await AllowAccess.findById(ids[i]);
+			let currentAccess = await AllowAccess.findById(ids[i]);
 			permitted.push(
 				/// @ts-ignore
-				contact.toObject({
+				currentAccess.toObject({
 					getters: true,
 					virtuals: true,
 				})
@@ -38,7 +38,10 @@ handler.post(async (req: NextApiRequest, res: NextApiResponse) => {
 		for (let i = 0; i < permitted.length; i++) {
 			const v = permitted[i];
 			/// @ts-ignore
-			await AllowAccess.findByIdAndRemove(v.id || v._id);
+			if (v.email !== "interchangefp@gmail.com") {
+				/// @ts-ignore
+				await AllowAccess.findByIdAndRemove(v.id || v._id);
+			}
 		}
 
 		let response = {
